Coverity IDs 126571, 126572, 126577
Fallout from the change from macros to inline functions, change a function
parameter from uint32 * to ushort * and remove the redundant type casts

Change a variable from unint32 to ushort

Finally, remove a redundant check for < 0 on an unsigned variable.

+++ base/gdevmpla.c
@@ -1770,7 +1770,7 @@ planar_to_chunky(gx_device_memory *mdev, int x, int y, int w, int h,
 
             for (pi = 0; pi < num_planes; ++pi) {
                 int plane_depth = mdev->planes[pi].depth;
-                uint value;
+                ushort value;
 
                 if (sample_load_next16(&value, &sptr[pi], &sbit[pi], plane_depth) < 0)
                     return_error(gs_error_rangecheck);
